requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
aiohappyeyeballs==2.4.6
    # via
    #   -r requirements/base.txt
    #   aiohttp
aiohttp==3.11.13
    # via
    #   -r requirements/base.txt
    #   litellm
aiosignal==1.3.2
    # via
    #   -r requirements/base.txt
    #   aiohttp
annotated-types==0.7.0
    # via
    #   -r requirements/base.txt
    #   pydantic
anyio==4.8.0
    # via
    #   -r requirements/base.txt
    #   httpx
    #   openai
appdirs==1.4.4
    # via
    #   -r requirements/base.txt
    #   fs
arrow==1.3.0
    # via cookiecutter
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
attrs==25.1.0
    # via
    #   -r requirements/base.txt
    #   aiohttp
    #   jsonschema
    #   referencing
binaryornot==0.4.4
    # via cookiecutter
boto3==1.37.1
    # via
    #   -r requirements/base.txt
    #   fs-s3fs
botocore==1.37.1
    # via
    #   -r requirements/base.txt
    #   boto3
    #   s3transfer
certifi==2025.1.31
    # via
    #   -r requirements/base.txt
    #   httpcore
    #   httpx
    #   requests
chardet==5.2.0
    # via binaryornot
charset-normalizer==3.4.1
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.8
    # via
    #   -r requirements/base.txt
    #   cookiecutter
    #   litellm
cookiecutter==2.6.0
    # via xblock-sdk
coverage[toml]==7.6.12
    # via pytest-cov
ddt==1.7.2
    # via -r requirements/test.in
distro==1.9.0
    # via
    #   -r requirements/base.txt
    #   openai
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-appconf
    #   django-statici18n
    #   edx-i18n-tools
    #   openedx-django-pyfs
    #   xblock-sdk
django-appconf==1.1.0
    # via
    #   -r requirements/base.txt
    #   django-statici18n
django-statici18n==2.6.0
    # via -r requirements/base.txt
edx-i18n-tools==1.6.3
    # via -r requirements/test.in
filelock==3.17.0
    # via
    #   -r requirements/base.txt
    #   huggingface-hub
frozenlist==1.5.0
    # via
    #   -r requirements/base.txt
    #   aiohttp
    #   aiosignal
fs==2.4.16
    # via
    #   -r requirements/base.txt
    #   fs-s3fs
    #   openedx-django-pyfs
    #   xblock
fs-s3fs==1.1.1
    # via
    #   -r requirements/base.txt
    #   openedx-django-pyfs
    #   xblock-sdk
fsspec==2025.2.0
    # via
    #   -r requirements/base.txt
    #   huggingface-hub
h11==0.14.0
    # via
    #   -r requirements/base.txt
    #   httpcore
httpcore==1.0.7
    # via
    #   -r requirements/base.txt
    #   httpx
httpx==0.28.1
    # via
    #   -r requirements/base.txt
    #   litellm
    #   openai
huggingface-hub==0.29.1
    # via
    #   -r requirements/base.txt
    #   tokenizers
idna==3.10
    # via
    #   -r requirements/base.txt
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.6.1
    # via
    #   -r requirements/base.txt
    #   litellm
iniconfig==2.0.0
    # via pytest
jinja2==3.1.5
    # via
    #   -r requirements/base.txt
    #   cookiecutter
    #   litellm
jiter==0.8.2
    # via
    #   -r requirements/base.txt
    #   openai
jmespath==1.0.1
    # via
    #   -r requirements/base.txt
    #   boto3
    #   botocore
jsonschema==4.23.0
    # via
    #   -r requirements/base.txt
    #   litellm
jsonschema-specifications==2024.10.1
    # via
    #   -r requirements/base.txt
    #   jsonschema
lazy==1.6
    # via
    #   -r requirements/base.txt
    #   xblock
litellm==1.61.17
    # via -r requirements/base.txt
lxml[html-clean]==5.3.1
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
    #   lxml-html-clean
    #   xblock
    #   xblock-sdk
lxml-html-clean==0.4.1
    # via lxml
mako==1.3.9
    # via
    #   -r requirements/base.txt
    #   xblock
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   -r requirements/base.txt
    #   jinja2
    #   mako
    #   xblock
mdurl==0.1.2
    # via markdown-it-py
mock==5.1.0
    # via -r requirements/test.in
multidict==6.1.0
    # via
    #   -r requirements/base.txt
    #   aiohttp
    #   yarl
openai==1.64.0
    # via
    #   -r requirements/base.txt
    #   litellm
openedx-django-pyfs==3.7.0
    # via
    #   -r requirements/base.txt
    #   -r requirements/test.in
    #   xblock
packaging==24.2
    # via
    #   -r requirements/base.txt
    #   huggingface-hub
    #   pytest
path==16.16.0
    # via edx-i18n-tools
pluggy==1.5.0
    # via pytest
polib==1.2.0
    # via edx-i18n-tools
propcache==0.3.0
    # via
    #   -r requirements/base.txt
    #   aiohttp
    #   yarl
pydantic==2.10.6
    # via
    #   -r requirements/base.txt
    #   litellm
    #   openai
pydantic-core==2.27.2
    # via
    #   -r requirements/base.txt
    #   pydantic
pygments==2.19.1
    # via rich
pypng==0.20220715.0
    # via xblock-sdk
pytest==8.3.4
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.0.0
    # via -r requirements/test.in
pytest-django==4.10.0
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/base.txt
    #   arrow
    #   botocore
    #   xblock
python-dotenv==1.0.1
    # via
    #   -r requirements/base.txt
    #   litellm
python-slugify==8.0.4
    # via cookiecutter
pytz==2025.1
    # via
    #   -r requirements/base.txt
    #   xblock
pyyaml==6.0.2
    # via
    #   -r requirements/base.txt
    #   cookiecutter
    #   edx-i18n-tools
    #   huggingface-hub
    #   xblock
referencing==0.36.2
    # via
    #   -r requirements/base.txt
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via
    #   -r requirements/base.txt
    #   tiktoken
requests==2.32.3
    # via
    #   -r requirements/base.txt
    #   cookiecutter
    #   huggingface-hub
    #   tiktoken
    #   xblock-sdk
rich==13.9.4
    # via cookiecutter
rpds-py==0.23.1
    # via
    #   -r requirements/base.txt
    #   jsonschema
    #   referencing
s3transfer==0.11.2
    # via
    #   -r requirements/base.txt
    #   boto3
simplejson==3.20.1
    # via
    #   -r requirements/base.txt
    #   xblock
    #   xblock-sdk
six==1.17.0
    # via
    #   -r requirements/base.txt
    #   fs
    #   fs-s3fs
    #   python-dateutil
sniffio==1.3.1
    # via
    #   -r requirements/base.txt
    #   anyio
    #   openai
sqlparse==0.5.3
    # via
    #   -r requirements/base.txt
    #   django
text-unidecode==1.3
    # via python-slugify
tiktoken==0.9.0
    # via
    #   -r requirements/base.txt
    #   litellm
tokenizers==0.21.0
    # via
    #   -r requirements/base.txt
    #   litellm
tqdm==4.67.1
    # via
    #   -r requirements/base.txt
    #   huggingface-hub
    #   openai
types-python-dateutil==2.9.0.20241206
    # via arrow
typing-extensions==4.12.2
    # via
    #   -r requirements/base.txt
    #   anyio
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
urllib3==2.2.3
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   botocore
    #   requests
web-fragments==2.2.0
    # via
    #   -r requirements/base.txt
    #   xblock
    #   xblock-sdk
webob==1.8.9
    # via
    #   -r requirements/base.txt
    #   xblock
    #   xblock-sdk
xblock[django]==4.0.1
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   xblock-sdk
xblock-sdk==0.12.0
    # via -r requirements/test.in
yarl==1.18.3
    # via
    #   -r requirements/base.txt
    #   aiohttp
zipp==3.21.0
    # via
    #   -r requirements/base.txt
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
